+++ lib/mixins/use_barlife.dart
@@ -8,12 +8,11 @@ mixin UseBarLife on Attackable {
   Color _backgroundColor = const Color(0xFF000000);
   Color _borderColor = const Color(0xFFFFFFFF);
   double _borderWidth = 2;
-  double _margin = 4;
   List<Color>? _colors;
-  Vector2? _offset;
+  Vector2? _barPosition;
   Vector2? _textOffset;
   BorderRadius _borderRadius = BorderRadius.zero;
-  BarLifePorition _barLifePosition = BarLifePorition.bottom;
+  BarLifeDrawPorition _barLifeDrawPosition = BarLifeDrawPorition.bottom;
   TextStyle? _textStyle;
   bool _showLifeText = true;
   ValueGeneratorComponent? _valueGenerator;
@@ -24,11 +23,10 @@ mixin UseBarLife on Attackable {
     Color? backgroundColor,
     Color? borderColor,
     double borderWidth = 2,
-    double margin = 4,
     List<Color>? colors,
     BorderRadius? borderRadius,
-    BarLifePorition barLifePosition = BarLifePorition.top,
-    Vector2? offset,
+    BarLifeDrawPorition barLifeDrawPosition = BarLifeDrawPorition.top,
+    Vector2? position,
     Vector2? textOffset,
     TextStyle? textStyle,
     bool showLifeText = true,
@@ -40,9 +38,8 @@ mixin UseBarLife on Attackable {
     _borderWidth = borderWidth;
     _colors = colors;
     _borderRadius = borderRadius ?? _borderRadius;
-    _barLifePosition = barLifePosition;
-    _margin = margin;
-    _offset = offset;
+    _barLifeDrawPosition = barLifeDrawPosition;
+    _barPosition = position;
     _textStyle = textStyle;
     _showLifeText = showLifeText;
     _textOffset = textOffset;
@@ -54,7 +51,7 @@ mixin UseBarLife on Attackable {
     add(
       barLife = BarLifeComponent(
         target: this,
-        position: _offset,
+        position: _barPosition,
         size: _barLifeSize ?? Vector2(width, 6),
         backgroundColor: _backgroundColor,
         borderColor: _borderColor,
@@ -63,8 +60,7 @@ mixin UseBarLife on Attackable {
         life: life,
         maxLife: maxLife,
         borderRadius: _borderRadius,
-        drawPosition: _barLifePosition,
-        margin: _margin,
+        drawPosition: _barLifeDrawPosition,
         textStyle: _textStyle,
         showLifeText: _showLifeText,
         textOffset: _textOffset,