+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+
+- `UseBarLife` improvements
 
 ## 3.0.0-beta.5