+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 3.0.0-beta.6
 
 - `UseBarLife` improvements
 - extension method `moveTowardsTarget` improvements